requirements.txt
websockets
stomper